Move diffusion coefficeints scalings into simulation support functions
