+++ docs/sim.rst
@@ -4,3 +4,8 @@ Simulation support
 ------------------
 .. automodule:: rbamlib.sim
 
+.. toctree::
+   :maxdepth: 2
+   :caption: Models
+
+   sim/dxx
