+++ rbamlib/models/__init__.py
@@ -1,16 +1,16 @@
 """
 The `models` package is a collection of scientific models for space physics research.
 
-Contains collection of empirical models of radial diffusion coefficient, lifetimes and local diffusion coefficients scaling, plasma densities, etc.
+Contains collection of empirical models of radial diffusion coefficient, lifetimes, plasma densities, etc.
 
 Packages:
+    - 'dip' : Dipole field model
     - 'lpp': Plasmapause location models
     - 'mp' : Magnetopause location models
-    - 'dxx' : Diffusion coefficient scaling models
-    - 'dll' : Radial diffusion models
-    - 'dip' : Dipole field model
     - 'mag' : Magnetic field related models (e.g., Tsyganenko coefficeints)
     - 'elec' : Electric field models
+    - 'dll' : Radial diffusion models    
+    - 'tau' : Lifetime models
     - 'pad' : Pitch-angle distribution models
     - 'ne' : Electron density models
     - 'ib' : Isotropic boundary models
