+++ rbamlib/sim/dxx/__init__.py
@@ -1,7 +1,7 @@
 """
-The `dxx` provides diffusion coefficient scaling and models
+The `sim.dxx` provides diffusion coefficient scaling
 
-Additional models and scaling functions are grouped into packages that correspond to specific waves, like `chorus` or `hiss`.
+Scaling functions are grouped into packages that correspond to specific waves, like `chorus` or `hiss`.
 
 Packages:
     - `chorus` : Models of chorus waves
