+++ tests/sim/dxx/test_lightning.py
@@ -1,9 +1,9 @@
 import unittest
 from tests.helpers import TestHelpers
-from rbamlib.models.dxx.vlf import S2011
+from rbamlib.sim.dxx.lightning import S2011
 
 
-class TestVlf(unittest.TestCase, TestHelpers):
+class TestLightning(unittest.TestCase, TestHelpers):
 
     def test_S2011(self):
         self.AssertBlank(S2011)
