+++ tests/sim/dxx/test_vlf.py
@@ -1,15 +1,12 @@
 import unittest
 from tests.helpers import TestHelpers
-from rbamlib.models.dxx.chorus import S2011, Z2019
+from rbamlib.sim.dxx.vlf import S2011
 
 
-class TestChorus(unittest.TestCase, TestHelpers):
+class TestVlf(unittest.TestCase, TestHelpers):
 
     def test_S2011(self):
         self.AssertBlank(S2011)
-
-    def test_Z2019(self):
-        self.AssertBlank(Z2019)
 
 
 if __name__ == '__main__':
